square is working

+++ include/VertexLayout.h
@@ -5,7 +5,7 @@
 class VertexLayout
 {
 public:
-    VertexLayout(size_t numAttributes) : m_numAttributes(numAttributes), m_attributeCount(0), m_readyForUse(false) {}
+    VertexLayout(unsigned numAttributes) : m_numAttributes(numAttributes), m_attributeCount(0), m_readyForUse(false) {}
     ~VertexLayout() = default;
 
     void SetLayout(size_t index, size_t size, size_t stride, size_t offset);
